[FIX] 온보딩 progress 수정 https://github.com/POLZZAK/POLZZAK-Android/issues/127
