--- app/src/main/java/com/polzzak_android/presentation/feature/onboarding/OnBoardingClickListener.kt
@@ -1,5 +1,0 @@
-package com.polzzak_android.presentation.feature.onboarding
-
-interface OnBoardingClickListener {
-    fun onNext()
-}
