+++ app/src/main/java/com/polzzak_android/presentation/feature/onboarding/item/OnBoardingPageItem.kt
@@ -1,20 +1,13 @@
 package com.polzzak_android.presentation.feature.onboarding.item
 
-import android.view.View
-import androidx.core.view.doOnPreDraw
-import androidx.core.view.isVisible
 import com.polzzak_android.R
 import com.polzzak_android.databinding.ItemOnBoardingPageBinding
 import com.polzzak_android.presentation.common.util.BindableItem
-import com.polzzak_android.presentation.common.util.toPx
-import com.polzzak_android.presentation.feature.onboarding.OnBoardingClickListener
 import com.polzzak_android.presentation.feature.onboarding.model.OnBoardingPageModel
 
 class OnBoardingPageItem(
     private val model: OnBoardingPageModel,
-    private val clickListener: OnBoardingClickListener
-) :
-    BindableItem<ItemOnBoardingPageBinding>() {
+) : BindableItem<ItemOnBoardingPageBinding>() {
     override val layoutRes: Int = R.layout.item_on_boarding_page
 
     override fun areItemsTheSame(other: BindableItem<*>): Boolean =
@@ -27,25 +20,6 @@ class OnBoardingPageItem(
         with(binding) {
             tvTitle.text = binding.root.context.getString(model.titleStringRes)
             tvContent.text = binding.root.context.getString(model.contentStringRes)
-            cpvProgress.maxCount = model.maxCount
-            cpvProgress.checkedCount = model.progress
-            val betweenMarginPx = PROGRESS_BETWEEN_MARGIN_DP.toPx(binding.root.context)
-            cpvProgress.betweenMarginsPx = betweenMarginPx
-            cpvProgress.doOnPreDraw {
-                cpvProgress.drawableWidthPx =
-                    (cpvProgress.width - (betweenMarginPx * model.maxCount - 1)) / model.maxCount
-            }
-            cpvProgress.isVisible = (model.progress < model.maxCount)
-            tvBtnStart.visibility =
-                if (model.progress == model.maxCount) View.VISIBLE else View.INVISIBLE
-            cpvProgress.checkedCount = position + 1
-            tvBtnStart.setOnClickListener {
-                clickListener.onNext()
-            }
         }
-    }
-
-    companion object {
-        private const val PROGRESS_BETWEEN_MARGIN_DP = 5
     }
 }
