+++ app/src/main/res/layout/fragment_on_boarding.xml
@@ -1,19 +1,63 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:background="@color/white">
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/vpPager"
             android:layout_width="0dp"
             android:layout_height="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintBottom_toTopOf="@id/clProgress"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
+
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/clProgress"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/vpPager">
+
+            <com.polzzak_android.presentation.component.CheckedProgressView
+                android:id="@+id/cpvProgress"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                app:checkedDrawable="@drawable/shape_rectangle_primary"
+                app:drawableHeight="4dp"
+                app:layout_constraintBottom_toBottomOf="@id/tvBtnStart"
+                app:layout_constraintEnd_toEndOf="@id/tvBtnStart"
+                app:layout_constraintStart_toStartOf="@id/tvBtnStart"
+                app:layout_constraintTop_toTopOf="@id/tvBtnStart"
+                app:uncheckedDrawable="@drawable/shape_rectangle_primary_150" />
+
+            <TextView
+                android:id="@+id/tvBtnStart"
+                style="@style/subtitle.16.600"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="16dp"
+                android:layout_marginBottom="36dp"
+                android:background="@drawable/selector_btn_rectangle_r8_enabled"
+                android:gravity="center"
+                android:paddingVertical="14dp"
+                android:text="@string/common_do_start"
+                android:textColor="@color/white"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                tools:isVisible="false" />
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
